pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "git-mcp-server"
version = "0.1.1RC1"
description = "API-first MCP server for Git repository workflows and branch-scoped file/directory tooling."
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
  { name = "Cloud-Dog, Viewdeck Engineering Limited" }
]
dependencies = [
    # Platform packages (from internal PyPI registry)
    "cloud_dog_config>=0.3.1",
    "cloud_dog_logging>=0.3.3",
    "cloud_dog_api_kit==0.4.1",
    "cloud_dog_idam>=0.2.0",
    "cloud_dog_db>=0.1.0",
    "cloud_dog_jobs>=0.3.0",
    "cloud_dog_cache>=0.1.0",
    "cloud_dog_storage>=0.1.4",
    # Core runtime
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.2,<3.0",
    "PyYAML>=6.0.1,<7.0",
    "typing-extensions>=4.10,<5.0",
    # Git operations
    "GitPython>=3.1.43,<4.0",
    # File tooling
    "filelock>=3.15,<4.0",
    "chardet>=5.2,<6.0",
    # Structured formats (edits + validation)
    "jsonschema>=4.22,<5.0",
    "lxml>=5.2,<6.0",
    "beautifulsoup4>=4.12,<5.0",
    "markdown-it-py>=3.0,<4.0",
    "ruamel.yaml>=0.18.6,<1.0",
    # Diff helpers
    "deepdiff>=7.0,<8.0",
    "unidiff>=0.7.5,<1.0",
    # Secrets / credentials
    "cryptography>=43.0,<47.0",
    "hvac>=2.3,<3.0",
    # ASGI server runtime for api_server.py / mcp_server.py entrypoints
    "uvicorn>=0.30,<1.0",
    "websockets>=12,<16",
]

[project.optional-dependencies]
dulwich = ["dulwich>=0.22.6,<1.0"]
dev = [
    "pytest>=8.2,<9.0",
    "pytest-cov>=5.0,<6.0",
    "pytest-asyncio>=0.23,<1.0",
    "ruff>=0.5,<1.0",
    "mypy>=1.10,<2.0",
    "httpx>=0.27,<1.0",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
ignore = ["UP017"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
disallow_untyped_decorators = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-p no:cloud_dog_config --basetemp=working/pytest-tmp"
markers = [
    "integration: integration tests that require live network/server behaviour",
]

[tool.hatch.build]
exclude = [
    "/.git",
    "/.mypy_cache",
    "/.pytest_cache",
    "/.ruff_cache",
    "/.venv",
    "/.tmp-tests",
    "/.pids",
    "/certs",
    "/data",
    "/dist",
    "/logs",
    "/working",
    "/vendor",
]

[tool.hatch.build.targets.wheel]
packages = ["src/git_mcp_server", "src/git_tools"]

[tool.hatch.build.targets.sdist]
only-include = [
    "src",
    "tests",
    "README.md",
    "pyproject.toml",
    "REQUIREMENTS.txt",
    "defaults.yaml",
    "ARCHITECTURE.md",
    "QUALITY-GATE.md",
    "RULES.md",
    "STANDARDS.md",
    "TESTS.md",
]

[project.urls]
"Source" = "https://github.com/cloud-dog-ai/git-mcp-server"
